Frontend build

+++ backend/app/models/payment.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Float
+from sqlalchemy import Column, String, Integer
 from app.database.database import Base
 
 class Payment(Base):
@@ -6,4 +6,4 @@ class Payment(Base):
 
     id = Column(String, primary_key=True)
     merchant_id = Column(String)
-    amount = Column(Float)
+    amount = Column(Integer)  # Financial amount in paise (integer)
