+++ backend/app/models/settlement.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Float
+from sqlalchemy import Column, String, Integer
 from app.database.database import Base
 
 class Settlement(Base):
@@ -6,4 +6,4 @@ class Settlement(Base):
 
     id = Column(String, primary_key=True)
     payment_id = Column(String)
-    amount = Column(Float)
+    amount = Column(Integer)  # Financial amount in paise (integer)
